--- a/python/sglang/srt/utils.py
+++ b/python/sglang/srt/utils.py
@@ -1148,15 +1148,19 @@ def point_to_point_pyobj(
     group: Optional[torch.distributed.ProcessGroup] = None,
     src: int = 0,
     dst: int = 1,
+    async_send: bool = False,
 ):
     """Send data from src to dst in group using DeviceToDevice communication."""
 
     if rank == src:
+        send_func = dist.send
+        if async_send:
+            send_func = dist.isend
         if len(data) == 0:
             tensor_size = torch.tensor(
                 [0], dtype=torch.long, device=torch.cuda.current_device()
             )
-            dist.send(tensor_size, dst=dst, group=group)
+            send_func(tensor_size, dst=dst, group=group)
         else:
             serialized_data = pickle.dumps(data)
             size = len(serialized_data)
@@ -1169,8 +1173,8 @@ def point_to_point_pyobj(
                 [size], dtype=torch.long, device=torch.cuda.current_device()
             )
 
-            dist.send(tensor_size, dst=dst, group=group)
-            dist.send(tensor_data, dst=dst, group=group)
+            send_func(tensor_size, dst=dst, group=group)
+            send_func(tensor_data, dst=dst, group=group)
         return data
 
     elif rank == dst: