## Motivation

pp send and recv could overlap

## Modifications

opt pp performance, enable pp send/recv overlap when pp size > 2

## my test result:

### Qwen3-235B-A22B 

TP8+PP4

start server command:

```
PYTHONUNBUFFERED=1 \
SGL_ENABLE_JIT_DEEPGEMM=0 \
python3 -m sglang.launch_server --model-path /qwen/Qwen3-235B-A22B --nnodes 4 --port 61001 --dist-init-addr 10.10.x.x:63001 --node-rank 0 --tp 8 --pp 4 --trust-remote-code --disable-radix-cache --cuda-graph-bs 1 2 4 8 16 32 64 128 256 512 768 896 1024 --mem-fraction-static 0.9 --max-running-requests 512 --chunked-prefill-size 16384 --context-length 16384 --attention-backend flashinfer --watchdog-timeout 3600
```

opt before:

```
# bench command
python3 -m sglang.bench_serving --port 61001 --dataset-name random-ids --num-prompts 128 --random-input-len 1000 --random-output-len 1000 --random-range-ratio 0.9 --disable-stream 

100%|██████████| 128/128 [00:47<00:00,  2.72it/s]

============ Serving Benchmark Result ============
Backend:                                 sglang    
Traffic request rate:                    inf       
Max request concurrency:                 not set   
Successful requests:                     128       
Benchmark duration (s):                  47.13     
Total input tokens:                      121127    
Total generated tokens:                  121703    
Total generated tokens (retokenized):    119159    
Request throughput (req/s):              2.72      
Input token throughput (tok/s):          2570.21   
Output token throughput (tok/s):         2582.43   
Total token throughput (tok/s):          5152.64   
Concurrency:                             122.87    
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   45239.21  
Median E2E Latency (ms):                 45290.44  
---------------Time to First Token----------------
Mean TTFT (ms):                          45239.23  
Median TTFT (ms):                        45290.46  
P99 TTFT (ms):                           47093.37  
---------------Inter-Token Latency----------------
Mean ITL (ms):                           0.00      
Median ITL (ms):                         0.00      
P95 ITL (ms):                            0.00      
P99 ITL (ms):                            0.00      
Max ITL (ms):                            0.00      
==================================================

# bench command
python3 -m sglang.bench_serving --port 61001 --dataset-name random-ids --num-prompts 256 --random-input-len 1000 --random-output-len 1000 --random-range-ratio 0.9 --disable-stream

100%|██████████| 256/256 [00:57<00:00,  4.46it/s]

============ Serving Benchmark Result ============
Backend:                                 sglang    
Traffic request rate:                    inf       
Max request concurrency:                 not set   
Successful requests:                     256       
Benchmark duration (s):                  57.34     
Total input tokens:                      242830    
Total generated tokens:                  243229    
Total generated tokens (retokenized):    241495    
Request throughput (req/s):              4.46      
Input token throughput (tok/s):          4234.78   
Output token throughput (tok/s):         4241.73   
Total token throughput (tok/s):          8476.51   
Concurrency:                             246.16    
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   55138.24  
Median E2E Latency (ms):                 55120.91  
---------------Time to First Token----------------
Mean TTFT (ms):                          55138.27  
Median TTFT (ms):                        55120.93  
P99 TTFT (ms):                           57303.99  
---------------Inter-Token Latency----------------
Mean ITL (ms):                           0.00      
Median ITL (ms):                         0.00      
P95 ITL (ms):                            0.00      
P99 ITL (ms):                            0.00      
Max ITL (ms):                            0.00      
==================================================
```

after:

```
# bench command
python3 -m sglang.bench_serving --port 61001 --dataset-name random-ids --num-prompts 128 --random-input-len 1000 --random-output-len 1000 --random-range-ratio 0.9 --disable-stream 

100%|██████████| 128/128 [00:38<00:00,  3.30it/s]

============ Serving Benchmark Result ============
Backend:                                 sglang    
Traffic request rate:                    inf       
Max request concurrency:                 not set   
Successful requests:                     128       
Benchmark duration (s):                  38.74     
Total input tokens:                      121127    
Total generated tokens:                  121703    
Total generated tokens (retokenized):    121837    
Request throughput (req/s):              3.30      
Input token throughput (tok/s):          3127.01   
Output token throughput (tok/s):         3141.88   
Total token throughput (tok/s):          6268.88   
Concurrency:                             123.26    
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   37300.16  
Median E2E Latency (ms):                 37367.44  
---------------Time to First Token----------------
Mean TTFT (ms):                          37300.18  
Median TTFT (ms):                        37367.47  
P99 TTFT (ms):                           38703.19  
---------------Inter-Token Latency----------------
Mean ITL (ms):                           0.00      
Median ITL (ms):                         0.00      
P95 ITL (ms):                            0.00      
P99 ITL (ms):                            0.00      
Max ITL (ms):                            0.00      
==================================================

# # bench command
100%|██████████| 256/256 [00:49<00:00,  5.17it/s]
python3 -m sglang.bench_serving --port 61001 --dataset-name random-ids --num-prompts 256 --random-input-len 1000 --random-output-len 1000 --random-range-ratio 0.9 --disable-stream

============ Serving Benchmark Result ============
Backend:                                 sglang    
Traffic request rate:                    inf       
Max request concurrency:                 not set   
Successful requests:                     256       
Benchmark duration (s):                  49.52     
Total input tokens:                      242830    
Total generated tokens:                  243229    
Total generated tokens (retokenized):    241541    
Request throughput (req/s):              5.17      
Input token throughput (tok/s):          4903.22   
Output token throughput (tok/s):         4911.28   
Total token throughput (tok/s):          9814.50   
Concurrency:                             246.85    
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   47753.80  
Median E2E Latency (ms):                 47750.40  
---------------Time to First Token----------------
Mean TTFT (ms):                          47753.82  
Median TTFT (ms):                        47750.42  
P99 TTFT (ms):                           49482.30  
---------------Inter-Token Latency----------------
Mean ITL (ms):                           0.00      
Median ITL (ms):                         0.00      
P95 ITL (ms):                            0.00      
P99 ITL (ms):                            0.00      
Max ITL (ms):                            0.00      
==================================================
```

#### profile:

before:

<img width="4200" height="1586" alt="image" src="https://github.com/user-attachments/assets/25a0b8ae-5093-48ff-9ed9-9c26de696aaf" />

after:

<img width="4190" height="1596" alt="image" src="https://github.com/user-attachments/assets/043a6eda-63c2-447c-876b-5cefe2ea4264" />


## DeepSeek-R1-Distill-Qwen-32B

tp8+pp4

before:

```
# bench command
python3 -m sglang.bench_serving --port 61001 --dataset-name random-ids --num-prompts 128 --random-input-len 1000 --random-output-len 1000 --random-range-ratio 0.9 --disable-stream 

100%|██████████| 128/128 [00:29<00:00,  4.30it/s]

============ Serving Benchmark Result ============
Backend:                                 sglang    
Traffic request rate:                    inf       
Max request concurrency:                 not set   
Successful requests:                     128       
Benchmark duration (s):                  29.79     
Total input tokens:                      121127    
Total generated tokens:                  121703    
Total generated tokens (retokenized):    119623    
Request throughput (req/s):              4.30      
Input token throughput (tok/s):          4066.43   
Output token throughput (tok/s):         4085.77   
Total token throughput (tok/s):          8152.20   
Concurrency:                             122.63    
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   28536.34  
Median E2E Latency (ms):                 28567.91  
---------------Time to First Token----------------
Mean TTFT (ms):                          28536.37  
Median TTFT (ms):                        28567.95  
P99 TTFT (ms):                           29759.67  
---------------Inter-Token Latency----------------
Mean ITL (ms):                           0.00      
Median ITL (ms):                         0.00      
P95 ITL (ms):                            0.00      
P99 ITL (ms):                            0.00      
Max ITL (ms):                            0.00      
==================================================
```

after

```
# bench command
python3 -m sglang.bench_serving --port 61001 --dataset-name random-ids --num-prompts 128 --random-input-len 1000 --random-output-len 1000 --random-range-ratio 0.9 --disable-stream 

100%|██████████| 128/128 [00:21<00:00,  5.98it/s]

============ Serving Benchmark Result ============
Backend:                                 sglang    
Traffic request rate:                    inf       
Max request concurrency:                 not set   
Successful requests:                     128       
Benchmark duration (s):                  21.42     
Total input tokens:                      121127    
Total generated tokens:                  121703    
Total generated tokens (retokenized):    119366    
Request throughput (req/s):              5.98      
Input token throughput (tok/s):          5655.69   
Output token throughput (tok/s):         5682.59   
Total token throughput (tok/s):          11338.28  
Concurrency:                             122.35    
----------------End-to-End Latency----------------
Mean E2E Latency (ms):                   20471.83  
Median E2E Latency (ms):                 20487.35  
---------------Time to First Token----------------
Mean TTFT (ms):                          20471.85  
Median TTFT (ms):                        20487.38  
P99 TTFT (ms):                           21380.34  
---------------Inter-Token Latency----------------
Mean ITL (ms):                           0.00      
Median ITL (ms):                         0.00      
P95 ITL (ms):                            0.00      
P99 ITL (ms):                            0.00      
Max ITL (ms):                            0.00      
==================================================
```
## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.