@Ying1123 2025-07-29
> > > > merged into #8346 Note: the performance of this implementation and the one in #8346 does not dominate each other. We chose #8346 because `isend` is a bit error-prone.
> > > 
> > > 
> > > isend is a bit error-prone？
> > > Specifically refers to the use of errors? Or somewhere else?
> > 
> > 
> > The current PR is good. I meant future developers might make mistakes on top of this easier:
> > https://docs.pytorch.org/docs/stable/distributed.html#torch.distributed.isend <img alt="Screenshot 2025-07-28 at 5 41 39 PM" width="875" height="226" src="https://private-user-images.githubusercontent.com/9302255/471757901-5b063185-0bfb-468b-b7ad-9a790abda787.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NTM3NTczMzAsIm5iZiI6MTc1Mzc1NzAzMCwicGF0aCI6Ii85MzAyMjU1LzQ3MTc1NzkwMS01YjA2MzE4NS0wYmZiLTQ2OGItYjdhZC05YTc5MGFiZGE3ODcucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDcyOSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTA3MjlUMDI0MzUwWiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9NGJiYzFlOTgxOTE5NmMxZGZlNzhmZDAzZjdiZWJmN2E2ZWRjNjM4ZDNkNWUyYTk5MTBjZjIzMzk2OTNiM2RiMiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.DLW3XCO1iQCGs2W2bG7hf584Wn3OppoXJ6qgJpnzUp8">
> 
> I don't deny that `isend` is an error-prone function. But do we have any other options to achieve the overlap of' send' and' recv'? I took a brief look at your PR, but you didn't realize the overlap of `send` and `recv`. I still don't understand, please help me explain.
> 
> The `isend` function is risky, but we just need to pay more attention when using it and write a good test case. I know that there is no good alternative to avoid using `isend` if we want to realize the overlap of `send` and `recv`.
> 
> At present, I can see that you are adding the number of micro-batch to achieve some overlap. This seems to be ok, too. Can you provide the test results similar to those in my PR?

Would you mind joining the SGL Slack (slack.sglang.ai)? We can have a direct chat there then. My account is sqy1415@gmail.com.